Add Mbed OS BSD socket adaptation layer
CHIP doesn't offer a platform abstraction for IP communication like it does for Bluetooth.
Instead the Inet layer is built on top of LwIP or the POSIX socket.

Mbed OS exposes neither to the user. LwIP may be use for some controllers but it is not exposed at the user level.
Instead Mbed OS offer an higher level C++ abstraction which is loosely based after BSD socket.
It allows integration with WiFi modules that just exposes an AT command interface and no bypass mode.

To integrate Mbed IP networking with Mbed OS, it has been chosen to develop an adaptation layer that exposes a
BSD socket compatible API. This avoid cluttering the Inet files with yet another networking option and retain compatibility
or Mbed OS targets not relying on LwIP.

This layer will eventually move out of CHIP to a separate repository or Mbed OS itself.